# src/api/routes.rs
//! Default unversioned Moneykeeper HTTP router and shared adapters.

use std::sync::Arc;

use axum::extract::{FromRequest, FromRequestParts, Request, State};
use axum::http::{Method, StatusCode, header::AUTHORIZATION, request::Parts};
use axum::middleware::{self, Next};
use axum::response::{IntoResponse, Response};
use axum::{Json, Router};
use jsonwebtoken::jwk::JwkSet;
use serde_json::json;
use uuid::Uuid;

use crate::api::jwt::verify_token;
use crate::api::middleware::AuthUser;
use crate::bootstrap::ContextFacades;
use crate::shared_kernel::UserId;

/// Composes all supporting and core context routes.
pub fn router(contexts: ContextFacades, jwks: Arc<JwkSet>) -> Router {
    let banking = contexts.banking.clone();
    let mail = contexts.mail.clone();
    let authenticated = Router::new()
        .merge(crate::contexts::portfolio::api::routes::router(
            contexts.portfolio,
        ))
        .merge(crate::contexts::sharing::api::routes::router(
            contexts.sharing,
            contexts.currencies.clone(),
        ))
        .merge(crate::contexts::ledger::api::routes::router(
            crate::api::state::LedgerApiState {
                ledger: contexts.ledger,
                currencies: contexts.currencies.clone(),
                banking: Some(banking.clone()),
            },
        ))
        .merge(crate::contexts::banking::api::routes::authenticated_router(
            banking.clone(),
        ))
        .merge(crate::contexts::reference_data::api::routes::router(
            contexts.currencies.clone(),
        ))
        .merge(crate::contexts::mail::api::routes::authenticated_router(
            mail.clone(),
        ))
        .merge(crate::contexts::recurring::api::routes::router(
            contexts.recurring,
        ))
        .merge(crate::contexts::reporting::api::routes::router(
            contexts.reporting,
        ))
        .merge(crate::contexts::loans::api::routes::router(
            contexts.loans,
            contexts.currencies.clone(),
        ))
        .merge(crate::contexts::classification::api::routes::router(
            contexts.categories,
        ))
        .merge(crate::contexts::preferences::api::routes::router(
            contexts.preferences,
            contexts.currencies,
        ))
        .layer(middleware::from_fn_with_state(
            AuthState { jwks },
            authenticate,
        ));
    crate::contexts::mail::api::routes::callback_router(mail)
        .merge(crate::contexts::banking::webhook_router(banking))
        .merge(authenticated)
        .layer(middleware::from_fn(reject_removed_legacy_mutations))
}

async fn reject_removed_legacy_mutations(request: Request, next: Next) -> Response {
    let path = request.uri().path().trim_matches('/');
    let segments: Vec<&str> = path.split('/').collect();
    let removed_hard_delete = request.method() == Method::DELETE
        && segments.len() == 2
        && matches!(segments.first().copied(), Some("accounts" | "transactions"));
    let removed_balance_setter = request.method() == Method::PATCH
        && segments.len() == 3
        && segments.first().copied() == Some("accounts")
        && segments.last().copied() == Some("balance");
    let missing_webhook_credential = path == "webhooks/monobank";
    let versioned_alias = path == "v2" || path.starts_with("v2/");
    if removed_hard_delete
        || removed_balance_setter
        || missing_webhook_credential
        || versioned_alias
    {
        return (StatusCode::NOT_FOUND, Json(json!({"error": "not_found"}))).into_response();
    }
    next.run(request).await
}

#[derive(Clone)]
struct AuthState {
    jwks: Arc<JwkSet>,
}

async fn authenticate(
    State(state): State<AuthState>,
    mut request: Request,
    next: Next,
) -> Result<Response, ApiError> {
    let token = request
        .headers()
        .get(AUTHORIZATION)
        .and_then(|value| value.to_str().ok())
        .and_then(|value| value.strip_prefix("Bearer "))
        .ok_or_else(ApiError::unauthorized)?;
    let claims = verify_token(token, &state.jwks).map_err(|_| ApiError::unauthorized())?;
    let user_id = Uuid::parse_str(&claims.sub).map_err(|_| ApiError::unauthorized())?;
    request.extensions_mut().insert(AuthUser(user_id));
    Ok(next.run(request).await)
}

/// The exact Moneykeeper method/path manifest used to validate OpenAPI parity.
pub const ROUTE_MANIFEST: &[(&str, &str)] = &[
    ("POST", "/portfolio-accounts"),
    ("GET", "/portfolio-accounts"),
    ("GET", "/portfolio-accounts/{id}"),
    ("PATCH", "/portfolio-accounts/{id}"),
    ("POST", "/portfolio-accounts/{id}/archive"),
    ("POST", "/portfolio-accounts/{id}/restore"),
    ("GET", "/portfolio-accounts/{id}/activity"),
    ("POST", "/instruments/ovdp"),
    ("GET", "/instruments"),
    ("GET", "/instruments/{id}"),
    ("POST", "/portfolio-transactions"),
    ("POST", "/portfolio-transactions/{id}/reversals"),
    ("GET", "/portfolio-positions"),
    ("POST", "/valuations"),
    ("GET", "/valuations"),
    ("GET", "/currencies"),
    ("GET", "/currencies/{code}"),
    ("POST", "/categories"),
    ("GET", "/categories"),
    ("GET", "/categories/{id}"),
    ("PATCH", "/categories/{id}"),
    ("POST", "/categories/{id}/archive"),
    ("POST", "/categories/{id}/restore"),
    ("GET", "/preferences"),
    ("PATCH", "/preferences"),
    ("POST", "/accounts"),
    ("GET", "/accounts"),
    ("GET", "/accounts/{id}"),
    ("PATCH", "/accounts/{id}"),
    ("POST", "/accounts/{id}/archive"),
    ("POST", "/accounts/{id}/restore"),
    ("GET", "/accounts/{id}/activity"),
    ("POST", "/transactions"),
    ("GET", "/transactions"),
    ("GET", "/transactions/{id}"),
    ("PATCH", "/transactions/{id}/annotation"),
    ("POST", "/transactions/{id}/reversals"),
    ("POST", "/transactions/{id}/replacements"),
    ("POST", "/transfers"),
    ("POST", "/accounts/{id}/balance-corrections"),
    ("GET", "/reconciliations"),
    ("GET", "/reconciliations/{id}"),
    ("POST", "/reconciliations/{id}/approve"),
    ("POST", "/reconciliations/{id}/dismiss"),
    ("POST", "/provider-connections/monobank"),
    ("GET", "/provider-connections"),
    ("GET", "/provider-connections/{id}"),
    ("POST", "/provider-connections/{id}/disconnect"),
    ("POST", "/provider-connections/{id}/credential-replacements"),
    ("POST", "/provider-connections/{id}/webhook-rotations"),
    ("GET", "/provider-connections/{id}/resources"),
    ("POST", "/provider-connections/{id}/resource-mappings"),
    (
        "POST",
        "/provider-connections/{id}/resource-mappings/{mapping_id}/deactivations",
    ),
    (
        "POST",
        "/provider-connections/{id}/resource-mappings/{mapping_id}/replacements",
    ),
    ("POST", "/provider-connections/{id}/sync-jobs"),
    ("GET", "/sync-jobs/{id}"),
    ("GET", "/provider-events/{id}"),
    ("GET", "/accounting-processes/{id}"),
    ("GET", "/balance-observations/{id}"),
    ("POST", "/me/email-connections/gmail/oauth/start"),
    ("GET", "/oauth/gmail/callback"),
    ("GET", "/me/email-connections"),
    ("GET", "/me/email-connections/{connection_id}/status"),
    ("POST", "/me/email-connections/{connection_id}/disconnect"),
    ("POST", "/me/email-connections/{connection_id}/resync"),
    ("GET", "/subscriptions"),
    ("GET", "/subscriptions/{subscription_id}"),
    ("PATCH", "/subscriptions/{subscription_id}"),
    ("GET", "/subscriptions/{subscription_id}/charges"),
    ("GET", "/subscriptions/forecast"),
    ("POST", "/subscription-charges/{charge_evidence_id}/matches"),
    (
        "POST",
        "/subscription-charges/{charge_evidence_id}/rejections",
    ),
    (
        "POST",
        "/subscription-charges/{charge_evidence_id}/matches/{match_id}/unmatches",
    ),
    ("GET", "/fx-rates"),
    ("GET", "/reports/balance-history"),
    ("GET", "/reports/cashflow"),
    ("GET", "/reports/spending"),
    ("GET", "/reports/liabilities"),
    ("GET", "/reports/reconciliations"),
    ("GET", "/reports/recurring"),
    ("GET", "/reports/net-worth"),
    ("GET", "/loans"),
    ("GET", "/loans/{id}"),
    ("GET", "/loans/{id}/term-revisions"),
    ("GET", "/loans/{id}/movements"),
    ("GET", "/loans/{id}/movements/{movement_id}"),
    ("POST", "/loans"),
    ("POST", "/loans/{id}/term-revisions"),
    ("POST", "/loans/{id}/closure"),
    ("POST", "/loans/{id}/disbursements"),
    ("POST", "/loans/{id}/repayments"),
    ("POST", "/loans/{id}/interest-accruals"),
    ("POST", "/loans/{id}/write-offs"),
    ("POST", "/loans/{id}/movements/{movement_id}/reversals"),
    ("POST", "/loans/{id}/movements/{movement_id}/replacements"),
    ("POST", "/contacts"),
    ("GET", "/contacts"),
    ("GET", "/contacts/{id}"),
    ("PATCH", "/contacts/{id}"),
    ("POST", "/contacts/{id}/archive"),
    ("POST", "/bill-splits"),
    ("GET", "/bill-splits"),
    ("GET", "/bill-splits/{id}"),
    ("POST", "/bill-splits/{id}/revisions"),
    ("POST", "/bill-splits/{id}/settlements"),
    (
        "POST",
        "/bill-splits/{id}/settlements/{settlement_id}/reversal",
    ),
    ("POST", "/bill-splits/{id}/cancellations"),
];

/// Authenticated tenant identity extracted from the existing auth boundary.
pub struct AuthenticatedUser(pub UserId);

impl<S> FromRequestParts<S> for AuthenticatedUser
where
    S: Send + Sync,
{
    type Rejection = ApiError;

    async fn from_request_parts(parts: &mut Parts, _state: &S) -> Result<Self, Self::Rejection> {
        parts
            .extensions
            .get::<AuthUser>()
            .map(|user| Self(UserId::new(user.0)))
            .ok_or_else(ApiError::unauthorized)
    }
}

/// JSON extractor that keeps every request failure on the stable JSON error
/// contract instead of leaking Axum's plain-text rejection responses.
pub struct ApiJson<T>(pub T);

impl<S, T> FromRequest<S> for ApiJson<T>
where
    S: Send + Sync,
    Json<T>: FromRequest<S>,
{
    type Rejection = ApiError;

    async fn from_request(request: Request, state: &S) -> Result<Self, Self::Rejection> {
        Json::<T>::from_request(request, state)
            .await
            .map(|Json(value)| Self(value))
            .map_err(|_| ApiError::bad_request("invalid JSON request"))
    }
}

/// Stable HTTP error translation for the Moneykeeper API.
#[derive(Debug)]
pub struct ApiError {
    status: StatusCode,
    message: &'static str,
}

impl ApiError {
    pub fn bad_request(message: &'static str) -> Self {
        Self {
            status: StatusCode::BAD_REQUEST,
            message,
        }
    }

    pub fn unauthorized() -> Self {
        Self {
            status: StatusCode::UNAUTHORIZED,
            message: "unauthorized",
        }
    }

    pub fn not_found(message: &'static str) -> Self {
        Self {
            status: StatusCode::NOT_FOUND,
            message,
        }
    }

    pub fn conflict(message: &'static str) -> Self {
        Self {
            status: StatusCode::CONFLICT,
            message,
        }
    }

    pub fn bad_gateway(message: &'static str) -> Self {
        Self {
            status: StatusCode::BAD_GATEWAY,
            message,
        }
    }

    pub fn internal() -> Self {
        Self {
            status: StatusCode::INTERNAL_SERVER_ERROR,
            message: "internal server error",
        }
    }
}

impl IntoResponse for ApiError {
    fn into_response(self) -> Response {
        (self.status, Json(json!({"error": self.message}))).into_response()
    }
}