# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moneykeeper",
    "version": "0.1.0-foundation",
    "description": "Moneykeeper context-first contract mounted by the default runtime at unversioned paths."
  },
  "paths": {
    "/portfolio-accounts": {
      "get": {"operationId":"listPortfolioAccounts","summary":"List portfolio accounts","description":"Returns active and archived manual portfolio accounts for the authenticated owner.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Portfolio accounts"},"401":{"$ref":"#/components/responses/Unauthorized"},"500":{"$ref":"#/components/responses/InternalServerError"}}},
      "post": {"operationId":"openPortfolioAccount","summary":"Open a portfolio account","description":"Creates a manual portfolio account at version one using a required idempotency key.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"201":{"description":"Portfolio account opened"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}
    },
    "/portfolio-accounts/{id}": {
      "get": {"operationId":"getPortfolioAccount","summary":"Get a portfolio account","description":"Returns one tenant-owned portfolio account and its optimistic version.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Portfolio account"},"404":{"$ref":"#/components/responses/NotFound"},"500":{"$ref":"#/components/responses/InternalServerError"}}},
      "patch": {"operationId":"renamePortfolioAccount","summary":"Rename a portfolio account","description":"Renames a portfolio account under an expected-version fence and idempotency key.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"200":{"description":"Portfolio account renamed"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}
    },
    "/portfolio-accounts/{id}/archive": {"post":{"operationId":"archivePortfolioAccount","summary":"Archive a portfolio account","description":"Archives an active portfolio account while preserving immutable activity and reversals.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"200":{"description":"Portfolio account archived"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/portfolio-accounts/{id}/restore": {"post":{"operationId":"restorePortfolioAccount","summary":"Restore a portfolio account","description":"Restores an archived portfolio account under an expected-version fence.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"200":{"description":"Portfolio account restored"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/portfolio-accounts/{id}/activity": {"get":{"operationId":"listPortfolioActivity","summary":"List portfolio activity","description":"Returns immutable transactions in deterministic effective-time and sequence order.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Portfolio activity"},"404":{"$ref":"#/components/responses/NotFound"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/instruments": {"get":{"operationId":"listPortfolioInstruments","summary":"List portfolio instruments","description":"Returns manual tenant-owned security definitions without consulting an external registry.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Portfolio instruments"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/instruments/{id}": {"get":{"operationId":"getPortfolioInstrument","summary":"Get a portfolio instrument","description":"Returns one manual instrument with explicit coupon and maturity terms.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Portfolio instrument"},"404":{"$ref":"#/components/responses/NotFound"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/instruments/ovdp": {"post":{"operationId":"createManualOvdpInstrument","summary":"Create a manual ОВДП instrument","description":"Records an ОВДП instrument with absolute face value, currency, dates, and explicit coupon knowledge.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"201":{"description":"Instrument created"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/portfolio-transactions": {"post":{"operationId":"recordPortfolioTransaction","summary":"Record immutable portfolio activity","description":"Records an opening, buy, sell, coupon, redemption, or correction with position-version fencing.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"201":{"description":"Portfolio transaction posted"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/portfolio-transactions/{id}/reversals": {"post":{"operationId":"reversePortfolioTransaction","summary":"Reverse portfolio activity","description":"Appends an exact compensating fact and coordinates cancellation or reversal of optional cash settlement.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"201":{"description":"Portfolio reversal posted"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/portfolio-positions": {"get":{"operationId":"listPortfolioPositions","summary":"List portfolio positions","description":"Returns exact quantity, known and unknown cost, realized result, and latest valuation state.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Portfolio positions"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/valuations": {
      "get":{"operationId":"listPortfolioValuations","summary":"List portfolio valuations","description":"Returns append-only absolute price and accrued-interest observations in deterministic order.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Portfolio valuations"},"500":{"$ref":"#/components/responses/InternalServerError"}}},
      "post":{"operationId":"recordPortfolioValuation","summary":"Record a portfolio valuation","description":"Appends an absolute per-instrument valuation and updates only rebuildable market-value projections.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"201":{"description":"Valuation recorded"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}
    },
    "/currencies": {
      "get": {
        "operationId": "listCurrencies",
        "summary": "List enabled currencies",
        "description": "Returns the enabled currencies recognized by Reference Data, ordered by currency code. This is a side-effect-free catalog query; disabled currencies are not included.",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Enabled currencies",
            "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Currency"}}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/currencies/{code}": {
      "get": {
        "operationId": "getCurrency",
        "summary": "Get an enabled currency",
        "description": "Returns one enabled currency by its canonical three-letter uppercase code. Missing and disabled currencies are both reported as not found, and the query does not change catalog state.",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "code", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Z]{3}$"}}],
        "responses": {
          "200": {"description": "Enabled currency", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Currency"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/categories": {
      "post": {
        "operationId": "createCategory",
        "summary": "Create a category",
        "description": "Creates an active category for the authenticated user at version 1. The name is trimmed, must contain 1 to 100 characters, and must be unique among that user's active categories under case-insensitive comparison.",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateCategory"}}}},
        "responses": {
          "201": {"description": "Created category", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Category"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      },
      "get": {
        "operationId": "listCategories",
        "summary": "List the user's categories",
        "description": "Returns all active and archived categories owned by the authenticated user. Archived categories remain visible so historical classifications are never hidden or deleted.",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "All active and archived user categories", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Category"}}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/categories/{id}": {
      "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
      "get": {
        "operationId": "getCategory",
        "summary": "Get a category",
        "description": "Returns an active or archived category owned by the authenticated user. A missing category and a category owned by another user are both reported as not found.",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Category", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Category"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      },
      "patch": {
        "operationId": "renameCategory",
        "summary": "Rename a category",
        "description": "Changes a category name when expected_version matches its current version. A real name change increments the version; replaying the same normalized name at the current version is idempotent. Stale versions and active-name collisions return a conflict.",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RenameCategory"}}}},
        "responses": {
          "200": {"description": "Renamed category", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Category"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/categories/{id}/archive": {
      "post": {
        "operationId": "archiveCategory",
        "summary": "Archive a category",
        "description": "Makes a category unavailable for new classifications while preserving its identity and historical use. The command requires the current expected_version and increments the version on the active-to-archived transition; repeating it against an already archived category at the current version is idempotent.",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExpectedVersion"}}}},
        "responses": {
          "200": {"description": "Archived category", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Category"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/categories/{id}/restore": {
      "post": {
        "operationId": "restoreCategory",
        "summary": "Restore an archived category",
        "description": "Makes an archived category active and available for new classifications again without losing its history or identity. The command requires the current expected_version and increments the version on the archived-to-active transition. It conflicts when the version is stale or another active category already has the same name under case-insensitive comparison; repeating it against an already active category at the current version is idempotent.",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExpectedVersion"}}}},
        "responses": {
          "200": {"description": "Restored category", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Category"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/preferences": {
      "get": {
        "operationId": "getPreferences",
        "summary": "Get effective preferences",
        "description": "Returns the authenticated user's effective preferences. If no row has been stored, the response uses base currency UAH with version 0 and persisted false without inserting data.",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Effective user preferences", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Preferences"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      },
      "patch": {
        "operationId": "setBaseCurrency",
        "summary": "Set the base currency",
        "description": "Sets the authenticated user's base currency to an enabled catalog currency using optimistic concurrency. expected_version 0 creates the first persisted preference at version 1; later changes must provide the current version and increment it on success.",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdatePreferences"}}}},
        "responses": {
          "200": {"description": "Updated preferences", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Preferences"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/accounts": {
      "post": {
        "operationId": "openLedgerAccount", "summary": "Open a Ledger account",
        "description": "Opens a tenant-owned manual account and records a non-zero opening balance through an immutable balanced journal. Requires Idempotency-Key.",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref":"#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OpenAccount"}}}},
        "responses": {"201": {"description": "Account opened", "content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountResult"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "409": {"$ref": "#/components/responses/Conflict"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      },
      "get": {
        "operationId": "listLedgerAccounts", "summary": "List Ledger accounts",
        "description": "Lists the authenticated user's visible active and archived accounts with exact signed and display balances.",
        "security": [{"bearerAuth": []}], "responses": {"200": {"description": "Ledger accounts", "content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LedgerAccount"}}}}}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      }
    },
    "/accounts/{id}": {
      "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
      "get": {
        "operationId": "getLedgerAccount", "summary": "Get a Ledger account",
        "description": "Returns one tenant-owned account; another user's identity is indistinguishable from a missing account.",
        "security": [{"bearerAuth": []}], "responses": {"200": {"description": "Ledger account", "content":{"application/json":{"schema":{"$ref":"#/components/schemas/LedgerAccount"}}}}, "404": {"$ref": "#/components/responses/NotFound"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      },
      "patch": {
        "operationId": "renameLedgerAccount", "summary": "Rename a Ledger account",
        "description": "Renames account metadata using expected_version and an Idempotency-Key without changing its currency or accounting facts.",
        "security": [{"bearerAuth": []}], "parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RenameAccount"}}}},
        "responses": {"200": {"description": "Account renamed", "content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountResult"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"$ref": "#/components/responses/Conflict"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      }
    },
    "/accounts/{id}/archive": {
      "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
      "post": {
        "operationId": "archiveLedgerAccount", "summary": "Archive a Ledger account",
        "description": "Archives an account at expected_version while preserving its full balance and immutable journal history.",
        "security": [{"bearerAuth": []}], "parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExpectedAccountVersion"}}}},
        "responses": {"200": {"description": "Account archived", "content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountResult"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "409": {"$ref": "#/components/responses/Conflict"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      }
    },
    "/accounts/{id}/restore": {
      "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
      "post": {
        "operationId": "restoreLedgerAccount", "summary": "Restore a Ledger account",
        "description": "Restores an archived account at expected_version without rewriting any historical accounting facts.",
        "security": [{"bearerAuth": []}], "parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExpectedAccountVersion"}}}},
        "responses": {"200": {"description": "Account restored", "content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountResult"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "409": {"$ref": "#/components/responses/Conflict"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      }
    },
    "/accounts/{id}/activity": {
      "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
      "get": {
        "operationId": "listLedgerAccountActivity", "summary": "List account activity",
        "description": "Lists immutable journal details with stable occurred-at and ledger-sequence cursor ordering.",
        "security": [{"bearerAuth": []}], "parameters":[{"$ref":"#/components/parameters/AfterOccurredAt"},{"$ref":"#/components/parameters/AfterSequence"},{"$ref":"#/components/parameters/ActivityLimit"}], "responses": {"200": {"description": "Account activity", "content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Journal"}}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      }
    },
...
    "/me/email-connections/gmail/oauth/start": {"post": {"operationId":"startGmailOauth","summary":"Start Gmail OAuth","description":"Persists encrypted single-use OAuth state and returns the provider redirect.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"OAuth redirect"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "parameters": {
      "IdempotencyKey": {"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":1,"maxLength":200}},
...
    },
    "responses": {
      "BadRequest": {"description": "Invalid request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Unauthorized": {"description": "Authentication required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "NotFound": {"description": "Resource not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Conflict": {"description": "Optimistic version or uniqueness conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "InternalServerError": {"description": "Supporting context unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
...
      "CreateCategory": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "kind"],
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 100},
          "kind": {"type": "string", "enum": ["income", "expense", "both"]}
        }
      },
...
    }
  }
}